pyproject.toml
[tool.poetry]
name = "autotest-api"
version = "0.1.0"
description = ""
authors = ["Alex Bezgliadnov <answer3ster@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
sqlalchemy = "^2.0.45"
alembic = "^1.17.2"
psycopg = "^3.3.2"
python-dotenv = "^1.2.1"
pydantic-settings = "^2.12.0"
pika = "^1.3.2"
redis = "^7.1.0"
playwright = "^1.57.0"
pytest = "^9.0.2"
pytest-playwright = "^0.7.2"
loguru = "^0.7.3"
minio = "^7.2.20"
pytest-asyncio = "^1.3.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.11"
mypy = "^1.19.1"
pytest = "^9.0.2"
httpx = "^0.28.1"
testcontainers = {extras = ["postgres"], version = "^4.14.0"}
psycopg2-binary = "^2.9.11"
pre-commit = "^4.5.1"
types-requests = "^2.32.4.20260107"
types-python-dateutil = "^2.9.0.20260124"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 100

# Include
# E/F — pyflakes + pycodestyle
# I — import sort
# UP — pyupgrade
# B — bugbear
# SIM — simplify
# C4 — comprehensions
# TCH — type-checking imports
# RUF — ruff-specific
lint.select = ["E", "F", "I", "UP", "B", "SIM", "C4", "TCH", "RUF"]

lint.ignore = [
    "E501", "B008"
]

exclude = ["tests", "alembic"]

fix = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
exclude = ["alembic"]

[tool.mypy]
python_version = "3.12"
strict = true

warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true

plugins = [
    "sqlalchemy.ext.mypy.plugin"
]

exclude = [
    "migrations/",
    "tests/",
    "alembic/"
]
